fix: reimplement idle state logic
implements idle state as another state in the FishState enum, and
includes related updates/additions to logic in FishSystemScript and FishScript.

also adjusts code in SlideBridge, as the game otherwise will not
compile. There's no guarantee that this script behaves as originally
intended, but it didn't seem to be finished at this time. It should now
set feedingIntensity properly in FishSystemScript, but I have not tested this.

Refs: #73