feat: change scoring so it doesn't decrease
Scoring won't decrease in the score and it won't be negative. Modified
scoring so it is also not dependent on the amount of fish in the merd.
Have also increased the time intervals the merds use between Full, Hungry and Dying stages. This still needs improvement since the slider doesn't work in the demo scene.

Refs: #98