fix: mute ocean sounds in office.
Moved the audio source a little to fix the problem. Not ideal at all,
but there is no way to mute or lower volume within a specified area as
far as I know. Lowering the volume based on distance from the source
like this does not work well when entering a building etc.

Refs: #147