feat: made FishWelfare look like FishFeed and connected with transition
Added the stuff from the EmptyScene so FishWelfare looks like FishFeed
on the outside. Added a free motorboat from the unity asset store and
connected the two scenes with a scene transition triggered on the
motorboat. Also added the starterTutorial in the FishFeed scene where
they learn the basic controls.

Refs: #162