fix: level select and tutorial popups.
Changed location and path for loading XML file, as the Assets folder and
so on does not exist in a build. Need to make a build to test if it
works.
Fixed the tutorial popups not behaving properly in advanced mode by
adding a loop that disables all tutorial entries in tutorial.cs

Refs: #210