fix: nauseating underwater bug
Was caused by unreferenced shader files. Not sure exactly which ones are
necessary, as building for Android tends to worsen/break shaders, making it
hard to see if effects are enabled properly or not. However, it is safe
to say that the one named "underwater" is. There is now a blue underwater hue and fog (probably nothing new), and the choppy effects on the left eye is gone. Still looks kinda bad because Android (I think), and the performance is, as already known, not very good.

Refs: #216