fix: fish LoD and button/hologram bloom
Enabled HDR, which enables bloom. Might have consequences for
performance, but is not noticable at the moment as there are other
factors (probably the fish system for the most part) affecting
performance more.
Also adjusted color of fish sprite to make the transition from 3D-model
to sprite less jarring. This might also mitigate the problem where fish
"dissapear" when looking up from the bottom of a cage.
Increased intensity of underwater shader, making it look closer to
PC/Unity.

Refs: #205