fix: did it properly
Made MerdCamera a child of CameraTrack, which fixes the problem. Also
attempted to draw the camera track in Unity GUI, replacing the useless
collider used for that purpose. It reacts to scaling etc, but is
probably not completely accurate as the math involved in MerdCamera is
confusing.
Lastly adjusted the scale of camera track to prevent camera from going
above water surface.

Refs: #205