fix: repair hologram and increase fish culling
The previous fixes broke the fish hologram. This is a workaround by
giving the user the option to increase/decrease the Z and Y-axis
speed/distance of the hologram camera.
Also increased the culling distance of fish, preventing them from
dissapearing. The main bug of dissapearing fish should now be fixed.

Refs: #205