feat: simplified and replaced 3 of 4 models
Used the Decimade Geometry tool in blender and reduced from around 1
million tris to 75-90k tris for 3 of the 4 models. Naturally introduces
tiny artefacts, which are not very noticable when playing. I had to do it myself
rather than using the new models, because this keeps the bones etc from
Blender when exporting (needed to make fish move and have physics).
Performance is greatly improved by doing this.

Refs: #230