feat: pointing controller will no longer manually refer to objects in the scene
change pointingcontroller and the dialogue script so that the object the npc will look at is set in the dialogue tree for each npc not in the pointing controller script. and remove duplicate npcs being stored in the npc list
note: this works in all scenes except fish feeding but i will look into it.