fix: make breaktask consistently
previously when moving out of collision box at the end of the breaktask
dialog it deletes/disables the anchor making BreakTask.cs stop running.
Which makes it impossible to start the next dialouge.
Sulution: Make it almost impossible by moving out of the collision box
making it fit the entire boat.