feat: implement unity events for dialogue tree changes
NpcTriggerDialogue.cs has a method which will be called when the dialogue changes. Which will keep track of the current and previous dialogue and passes the previous dialogue to the ResetDialogue method. When the player is returned to the previous dialogue in the laboratory it will now return to the exact element in the tree

Refs: #737

Co-authored-by: eamathie <69533149+eamathie@users.noreply.github.com>