feat: Add toggling of AI functionality
Add an NPC at the ReceptionOutdoor Spawn point that will ask the user about consenting to AI features. Add loading scene when toggling AI features on and off. Make method LoadSceneWithLoadingScreen public so that it can be accessed in ConsentForm.cs. Add a condition in method SpawnNPC in NPCSpawner.cs to listen to toggling state.